+++ app/src/main/java/com/meezzi/fingerchoice/ui/explore/ExploreRestaurantFragment.kt
@@ -1,4 +1,4 @@
-package com.meezzi.fingerchoice.explore
+package com.meezzi.fingerchoice.ui.explore
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
