+++ app/src/main/res/layout/activity_sign_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/orange"
-    tools:context=".signin.SignInActivity">
+    tools:context=".ui.signin.SignInActivity">
 
     <TextView
         android:id="@+id/textView"
