+++ app/src/main/res/layout/fragment_explore_restaurant.xml
@@ -4,6 +4,6 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".explore.ExploreRestaurantFragment">
+    tools:context=".ui.explore.ExploreRestaurantFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
