+++ app/src/main/res/layout/fragment_profile.xml
@@ -4,6 +4,6 @@
     android:id="@+id/frameLayout3"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".profile.ProfileFragment">
+    tools:context=".ui.profile.ProfileFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
