+++ app/src/main/res/navigation/home_graph.xml
@@ -7,13 +7,13 @@
 
     <fragment
         android:id="@+id/navigation_explore_restaurant"
-        android:name="com.meezzi.fingerchoice.explore.ExploreRestaurantFragment"
+        android:name="com.meezzi.fingerchoice.ui.explore.ExploreRestaurantFragment"
         android:label="ExploreRestaurantFragment"
         tools:layout="@layout/fragment_explore_restaurant" />
 
     <fragment
         android:id="@+id/navigation_map"
-        android:name="com.meezzi.fingerchoice.map.MapFragment"
+        android:name="com.meezzi.fingerchoice.ui.map.MapFragment"
         android:label="MapFragment"
         tools:layout="@layout/fragment_map" >
         <action
@@ -23,13 +23,13 @@
 
     <fragment
         android:id="@+id/navigation_saved_restaurant"
-        android:name="com.meezzi.fingerchoice.save.SavedRestaurantFragment"
+        android:name="com.meezzi.fingerchoice.ui.save.SavedRestaurantFragment"
         android:label="SavedRestaurantFragment"
         tools:layout="@layout/fragment_saved_restaurant" />
 
     <fragment
         android:id="@+id/navigation_profile"
-        android:name="com.meezzi.fingerchoice.profile.ProfileFragment"
+        android:name="com.meezzi.fingerchoice.ui.profile.ProfileFragment"
         android:label="ProfileFragment"
         tools:layout="@layout/fragment_profile" />
